requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cron-descriptor==1.4.3
cycler==0.12.1
Django==5.0.7
django-celery-beat==2.6.0
django-timezone-field==7.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
fonttools==4.53.1
kiwisolver==1.4.5
kombu==5.3.7
matplotlib==3.9.1
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
pyparsing==3.1.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.0.7
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
